+++ apps/tangle-dapp/data/ServiceTables/index.ts
@@ -1,2 +1,1 @@
 export { default as getActiveServices } from './getActiveServices';
-export { default as useActiveServicesByValidator } from './useActiveServicesByValidator';
